Base bot configuration, HTTP Client, Spotify API
Created a simple HTTP Client class with get/post methods.
Also created a Spotify API class with basic methods:
    - get_token() for requesting a token
    - __get_auth_header() returns header object for
      every request to the Spotify API
    - get_genres() requests the API for a music genres list
    - search() reusable search request for searching artists or tracks,
      search type based on provided enum value. Returns a list of search
      result objects { name, id }
    - get_recs() requests to generate recommendations based on provided
      seeds and a limit of needed tracks

Implemented a utility function to reset and set the list of bot commands
(from the config file) when the bot starts.
Created files for base handlers: general and for handling the start command.